python/trace-filtering/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation==0.50b0
opentelemetry-distro==0.50b0
